Adding TL_chart_add_data tests

+++ sources/indicators/ma.c
@@ -7,7 +7,7 @@
 extern int g_total_charts_allocated;
 extern TL_CHART *g_chart;
 extern TL_CHART_DATA **g_chart_data;
-extern size_t *g_chart_data_length;
+extern unsigned int *g_chart_data_length;
 
 TL_STATUS TL_ma(TL_CHART chart, TL_OHLC ohlc_type, unsigned int period, double *output)
 {
